+++ config/passport.ts
@@ -1,8 +1,8 @@
-const LocalStrategy = require("passport-local").Strategy;
-const mongoose = require("mongoose");
-const User = require("../models/User");
+import { Strategy as LocalStrategy } from "passport-local";
+import mongoose from "mongoose"
+import User from "../models/User"
 
-module.exports = function (passport) {
+export default function (passport) {
   passport.use(
     new LocalStrategy({ usernameField: "email" }, (email, password, done) => {
       User.findOne({ email: email.toLowerCase() }, (err, user) => {
