+++ middleware/cloudinary.ts
@@ -1,4 +1,4 @@
-const cloudinary = require("cloudinary").v2;
+import { v2 as cloudinary } from 'cloudinary'
 
 require("dotenv").config({ path: "./config/.env" });
 
@@ -8,4 +8,4 @@ cloudinary.config({
   api_secret: process.env.API_SECRET,
 });
 
-module.exports = cloudinary;
+export default cloudinary;
