+++ middleware/notifications.ts
@@ -1,15 +1,15 @@
-const Notification = require('../models/Notification');
-const User = require('../models/User');
+import User from '../models/User'
 
-
-async function notifications(req, res, next) {
+async function notifications(req:any, res:any, next:any) {
     if (!req.user) {
       return next();
     }
   
     try {
       const count = await User.findById(req.user._id); 
+      if(count != null){
       res.locals.notificationsCount = count.unreadCount || 0; // Initialize count to zero if undefined or null
+      }
       return next();
     } catch (err) {
       console.error(err);
@@ -17,4 +17,4 @@ async function notifications(req, res, next) {
     }
   }
   
-  module.exports = notifications;
+export default notifications;
