+++ models/Notification.ts
@@ -1,4 +1,4 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose'
 
 
 const NotificationSchema = new mongoose.Schema({
@@ -40,5 +40,6 @@ const NotificationSchema = new mongoose.Schema({
 });
 
 const Notification = mongoose.model('Notification', NotificationSchema);
-module.exports = Notification;
+
+export default Notification;
