+++ models/Relationship.ts
@@ -1,8 +1,8 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose'
 
 const RelationshipSchema = new mongoose.Schema({
   follower: { type: mongoose.Schema.Types.ObjectId, ref: 'User' },
   following: { type: mongoose.Schema.Types.ObjectId, ref: 'User' }
 });
 
-module.exports = mongoose.model('Relationship', RelationshipSchema);
+export default mongoose.model('Relationship', RelationshipSchema);
