+++ routes/posts.ts
@@ -1,14 +1,14 @@
-const express = require("express");
+import express from "express"
 const router = express.Router();
-const upload = require("../middleware/multer");
-const postsController = require("../controllers/posts");
-const { ensureAuth, ensureGuest } = require("../middleware/auth");
+import upload from "../middleware/multer"
+import postsController from "../controllers/posts"
+import authMiddleware from '../middleware/auth';
 
 
 //Post Routes - simplified for now
-router.get("/:id", ensureAuth, postsController.getPost);
+router.get("/:id", authMiddleware.ensureAuth, postsController.getPost);
 
-router.get("/:id/likes", ensureAuth, postsController.getLikes);
+router.get("/:id/likes", authMiddleware.ensureAuth, postsController.getLikes);
 
 router.post("/createPost", upload.single("media"), postsController.createPost);
 
@@ -18,4 +18,4 @@ router.post("/likePost/:id", postsController.likePost);
 
 router.delete("/deletePost/:id", postsController.deletePost);
 
-module.exports = router;
+export default router;
