--- routes/prs.js
@@ -1,16 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const upload = require("../middleware/multer");
-const prsController = require("../controllers/prs");
-const { ensureAuth, ensureGuest } = require("../middleware/auth");
-
-
-//Pr Routes - simplified for now
-
-router.post("/createPR", upload.single("media"), prsController.createPR);
-
-router.post("/likePR/:id", prsController.likePR);
-
-router.delete("/deletePR/:id", prsController.deletePR);
-
-module.exports = router;
